refactor: rename browser directory to browser-control
