+++ packages/tasks/src/task/browser-control/register.ts
