+++ packages/tasks/src/task/browser-control/tasks/BrowserClickTask.ts
