+++ packages/tasks/src/task/browser-control/tasks/BrowserNewTabTask.ts
